feat(repo): update workflows to support multiple environments (#121)
### Production
https://trailblazers.taiko.xyz/
Updated via releases on GH with the format: `trailblazer-ui-v` e.g.
`trailblazer-ui-v1.2.0`

**Naming convention for releases:** 
`Season.Trail.Fix`, so right now we could be `1.4.0.`, season 1,
trail/week 4 and initial release. If we need to make a new deployment
during the trail it will be `1.4.1` and so on.

**Hotfix flow:**

1. merge fix into `main`
2. cherry pick into release tag
3. trigger GH action on the release tag to deploy to prod again

### Staging
https://stagingtrailblazer-ui-taikoxyz.vercel.app/
Updated via merges to `main` branch

Must be stable and ready to go out via a release

###  QA
https://qatrailblazerui-taikoxyz.vercel.app/
Updated via merges to `qa` branch

May be unstable and overwritten by other devs (announce it quickly on
discord).

---------

Co-authored-by: Karim <162329697+kimo-ice@users.noreply.github.com>
Co-authored-by: Karim <karim@taiko.xyz>